sw: notebook: Update the wdt logic for subsystem restart
Earlier the wdt recovery would result in system restart or POR.
This allowed to disable all the actions permanently, as it was
confirm that the flow won't return for the next action.

Now the wdt recovery can be subsystem restart, disable the clickable
action only till the subsystem is alive (e.g. As soon as RPU subsystem
is dead, return the controls to normal behavior).

Signed-off-by: Pallav Joshi <pallav.joshi@xilinx.com>